Add files via upload
Description:
The project is built using Assembly Language in DOSBox, an x86 emulator with DOS. The game resembles Mario-Cart with a few design exceptions. The gui involves a Ginger-Bread-Man made using pixel-art with different coloured pixels stored in an array. There are clouds, skulls and candy canes as game prop objects. There are three levels to the game where by at the last level, a witch is used as an opponent which throws potions at the main character. Any collision with the potion results in losing lives until all 3 three
lives have been depleted .

Main Features:
1. Making animated characters using pixel colours stored in an array
2. Transitions between levels
3. The movement of the main character through arrow keys and space bar to jump over obstacles
3. Keeping track of the scores , incremented as each obstacle (candy cane) is overcome.
4. Keeping track of lives lost
5. Collision detection with enemies (ghosts and witch potion)